added code to automatically backup data

+++ widgets/viewer.py
@@ -5,16 +5,19 @@
 #
 import os
 import pickle
+import warnings
 from collections import defaultdict
 from copy import copy, deepcopy
 from datetime import datetime
+from pathlib import Path
 from random import sample as smp
+from typing import Tuple
 
 import numpy as np
 import pandas as pd
 from PIL import Image
 from PyQt5.Qt import pyqtSignal
-from PyQt5.QtCore import Qt, QTimer
+from PyQt5.QtCore import QTimer
 from PyQt5.QtGui import QFontMetrics
 from PyQt5.QtWidgets import (
     QFileDialog,
@@ -30,6 +33,7 @@ from utils import BoxLoader, Segmentation, get_2d_files, read_skeleton
 from .actionbar import Bar
 from .canvas import VideoCanvas
 from .console import Console
+from .core.annotations import save_annotations
 from .dialog import (
     AssessmentDialog,
     CatDialog,
@@ -735,16 +739,18 @@ class Viewer(QWidget):
         inv = {v: k for k, v in self.catDict[key].items()}
         return inv
 
-    def save(self, event=None, verbose=True, new_file=False, ask=False):
-        if ask:
-            msg = QMessageBox()
-            msg.setText("Save the current annotation?")
-            msg.setStandardButtons(QMessageBox.Yes | QMessageBox.No)
-            reply = msg.exec_()
-            if reply == QMessageBox.No:
-                return
-        self.update_labels()
+    def export_annotation_data(self) -> Tuple[dict, list, list, list]:
+        """
+        Updates the annotation times by calling self.update_labels(), and then prepares
+        the annotation data to be exported
 
+        Returns:
+            metadata
+            category_labels
+            animals
+            times
+        """
+        self.update_labels()
         times = deepcopy(self.times)
         for i, ind_list in enumerate(times):
             for j, cat_list in enumerate(ind_list):
@@ -758,6 +764,7 @@ class Viewer(QWidget):
         neg_classes = self.settings["hard_negative_classes"]
         if neg_classes == "all":
             neg_classes = [x for x in cat_labels if x not in self.invisible_actions]
+
         if neg_classes is not None and self.al_points is not None:
             for neg_class in neg_classes:
                 neg_ind = cat_labels.index(neg_class)
@@ -778,24 +785,46 @@ class Viewer(QWidget):
                     times[i].append([])
                     for start, end in zip(idx[:-1:2], idx[1::2]):
                         times[i][-1].append([start, end, 0])
+
         for j, a in enumerate(self.neg_actions):
             if self.neg_actions[j] not in cat_labels:
                 cat_labels.append(self.neg_actions[j])
                 for i in range(len(times)):
                     times[i].append(self.negative_times[i][j])
+
         for j, a in enumerate(self.unknown_actions):
             if self.unknown_actions[j] not in cat_labels:
                 cat_labels.append(self.unknown_actions[j])
                 for i in range(len(times)):
                     times[i].append(self.unknown_times[i][j])
+
+        metadata = {
+            "datetime": datetime.now().strftime("%d/%m/%Y %H:%M:%S"),
+            "annotator": self.settings["annotator"],
+            "length": self.video_len(),
+            "video_file": self.filenames[0],
+            "skeleton_files": self.settings["skeleton_files"],
+        }
+        return metadata, cat_labels, self.animals, times
+
+    def save(self, event=None, verbose=True, new_file=False, ask=False):
+        if ask:
+            msg = QMessageBox()
+            msg.setText("Save the current annotations?")
+            msg.setStandardButtons(QMessageBox.Yes | QMessageBox.No)
+            reply = msg.exec_()
+            if reply == QMessageBox.No:
+                return
+
+        metadata, cat_labels, animals, times = self.export_annotation_data()
         if self.output_file is None or new_file:
             self.output_file = QFileDialog.getSaveFileName(self, "Save file")[0]
             if len(self.output_file) == 0:
                 self.output_file = None
                 return False
 
+        # TODO: WHY IS THE LAST ACTION CHOICE SAVED HERE
         with open("../last_action_choice.pickle", "wb") as f:
-            # TODO: WHY IS THE LAST ACTION CHOICE SET HERE
             loaded_shortcuts = defaultdict(lambda: {})
             for k in self.catDict:
                 for i, sc in self.shortCutInv(k).items():
@@ -810,18 +839,22 @@ class Viewer(QWidget):
                 f,
             )
 
-        metadata = {}
-        metadata["datetime"] = datetime.now().strftime("%d/%m/%Y %H:%M:%S")
-        metadata["annotator"] = self.settings["annotator"]
-        metadata["length"] = self.video_len()
-        metadata["video_file"] = self.filenames[0]
-        metadata["skeleton_files"] = self.settings["skeleton_files"]
+        # TODO: SAVING
+        try:
+            save_annotations(
+                output_path=Path(self.output_file),
+                metadata=metadata,
+                animals=animals,
+                cat_labels=cat_labels,
+                times=times,
+                human_readable=True,
+                overwrite=True,
+            )
+            if verbose:
+                self.show_warning("Saved successfully!")
+        except IOError as err:
+            warnings.warn(f"Failed to save annotation data: {err}")
 
-        with open(self.output_file, "wb") as f:
-            # TODO: SAVING
-            pickle.dump((metadata, cat_labels, self.animals, times), f)
-        if verbose:
-            self.show_warning("Saved successfully!")
         return True
 
     def show_warning(self, message):
